common: Move container type into separate package
Also:
Including an entire package just to retrieve the definition of the
ContainerName label is pointless as it will never ever change again.

Size reduction:
 cilium            27M => 17M
 cilium-agent      37M => 37M
 cilium-docker     26M => 16M
 cilium-cni        26M => 16M
