+++ NEWS.rst
@@ -14,6 +14,7 @@ Features
 --------
 
 - Improved logging readability (`GH #499 <https://github.com/cilium/cilium/pull/499>`_)
+- Reduced size of cilium binary from 27M to 17M
 
 0.8.0
 =====
